Support automated customization of templates, scripts, tiles

+++ removeBlob.sh
@@ -16,12 +16,17 @@ fi
 blobFileName=$1
 
 shaFileId=`grep -A2 $blobFileName config/blobs.yml | grep sha  | awk '{print $2}' `
-echo "Deleting the blob entry: $blobFileName, cancel to stop"
-sleep 5
-rm .blobs/$shaFileId
-find blobs -name $blobFileName | xargs rm 
-sed -i.bak "/$blobFileName/,/size: / { d; }" config/blobs.yml
-rm config/blobs.yml.bak
+if [ "$shaFileId" != "" ]; then
+  echo "Deleting the blob entry: $blobFileName, cancel to stop"
+  sleep 5
+  rm .blobs/$shaFileId
+  find blobs -name $blobFileName | xargs rm 
+  sed -i.bak "/$blobFileName/,/size: / { d; }" config/blobs.yml
+  rm config/blobs.yml.bak
+
+  specFile=`grep $blobFileName packages/*/spec | awk -F ":" '{print $1}' `
+  sed -i.bak "/$blobFileName/ { d; }" $specFile 
+fi
 
 # The previous edit might leave the blobs.yml incomplete with just '---' without the '{}'
 # Recreate a pristine blobs.yml that has '--- {}'
